+++ src/main/java/net/qiuyu/horrorcooked9/register/ModEffects.java
@@ -11,17 +11,17 @@ import net.qiuyu.horrorcooked9.effects.DiarrheaEffect;
 import net.qiuyu.horrorcooked9.effects.PineapplePowerIEffect;
 
 public class ModEffects {
-    public static final DeferredRegister<MobEffect> ITEMS =
+    public static final DeferredRegister<MobEffect> MOB_EFFECTS =
             DeferredRegister.create(ForgeRegistries.MOB_EFFECTS, HorrorCooked9.MODID);
 
     public static final RegistryObject<MobEffect> CAPTAIN_INSPIRATION =
-            ITEMS.register("captain_inspiration", CaptainInspiration::new);
+            MOB_EFFECTS.register("captain_inspiration", CaptainInspiration::new);
     public static final RegistryObject<MobEffect> DIARRHEA =
-            ITEMS.register("diarrhea", DiarrheaEffect::new);
+            MOB_EFFECTS.register("diarrhea", DiarrheaEffect::new);
     public static final RegistryObject<MobEffect> PINEAPPLE_POWER_I =
-            ITEMS.register("pineapple_power_i", PineapplePowerIEffect::new);
+            MOB_EFFECTS.register("pineapple_power_i", PineapplePowerIEffect::new);
 
     public static void register(IEventBus eventBus) {
-        ITEMS.register(eventBus);
+        MOB_EFFECTS.register(eventBus);
     }
 }